backend/users/contactserializers.py
from rest_framework import serializers
from .contactmodels import ContactMessage

class ContactMessageSerializer(serializers.ModelSerializer):
    class Meta:
        model = ContactMessage
        fields = '__all__'
