requirements.txt
asgiref==3.9.1
Django==3.1.12
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.7.0
pillow==11.3.0
PyJWT==2.10.1
pymongo==4.14.1
pytz==2025.2
setuptools==80.9.0
sqlparse==0.2.4
tzdata==2025.2
